Update README.md
I made a windows app to provide a faster way to edit the needed files to make a new mod. 

It takes a few inputs, clones the directory using the new mods name and renames the relevant fields and file names. 

Check it out [here](https://github.com/Jellypowered/PrepareNewMod)